Rewrote base code, added mouse avoidance, and stress levels
The entire system was rather bad, so I reworked it and rewrote it (placing the old code in v1.gd, of course),
Afterwards, I got the cat to high-tail it if the window interacts with the mouse.  And, the feature I'm most excited about: stress! The cat gets stressed if it bumps into the edges or the mouse, increasing its speed.

Otherwise, I've also added more comments and plan to add an energy feature.